requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
discord.py==2.4.0
websockets==12.0
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pip==24.3.1
PyNaCl==1.5.0
aiofiles==24.1.0
libsql-client==0.9.2